requirements.txt
matplotlib==3.6.2
numpy==1.23.5
opencv_python==4.7.0.68
pandas==1.5.2
Pillow==9.4.0
keras==2.11.0
python_dateutil==2.8.2
requests==2.25.1
scikit_learn==1.2.1
seaborn==0.11.2
tensorflow==2.11.0
tensorflow_intel==2.11.0
xgboost==1.7.1

